Initial corebook loader upload.
xml -> json converter for the corebook. Note that the script requires a data file which has been converted to utf-8, rather than the utf-16 which the Pokerole corebook data comes in by default.

Currently only processes Pokémon - and by choice due to this still being in the testing phase, only Bulbasaur at that - but the script should also work for moves and abilities.